Implement Kafka setup: Product as consumer, Order as producer

+++ Product/Product.Domain/Entity/Product.cs
@@ -5,7 +5,7 @@
         public string Name { get; set; }
         public string Description { get; set; }
         public decimal Price { get; set; }
-        public int QuantityAvailable { get; set; }
+        public decimal QuantityAvailable { get; set; }
         public long CategoryId { get; set; }
         public bool IsActive { get; set; }
         public Category Category { get; set; }
